Further MPMC-Improvements
First I renamed the Functions to enqueue data on the bounded Versions, as they can also
fail which is why I prefixed them with "try" to signal that this might very well fail.

I then also added more comments and documentation to the Code in general to help with
understanding the Code from both the maintainers view as well as from the consumer
perspective, trying to use the crate.