Updated Safety
Made the retire function on the Hazard-Pointer Domain unsafe, because there are some
restrictions that the type-system cannot prove but they are still very much needed for it
to still be correct. This also meant documenting the Safety restrictions for it on the
function which actually helps a lot more with making sure its used corretly.

The other changes were mostly about adapting the use-cases of the retire function in the
rest of the codebase and adding the right safety comments to them