Added more tests and packed things more together
Made more of the constructors const and using that to move more of the Allocator data into the Allocator struct and moving all
the shared Data into a normal static instead of a lazy_static, which helps a bit with performance and cuts down on dependencies.
Then also improved some of the benchmarks as well as adding a couple more minor tests.