Started work on no_std support, minor docs improvements and also fixed some features
First of fixed a small Problems of the unbounded mpmc queue which relies on hazard-pointers, but hazard-pointers were not
enabled when enabling the queues feature on its own, now the unbounded-mpmc queue is only available when the hazard-pointer
feature is enabled.

Following inline with the features stuff, also now added the feature annotation to the items in the Docs, which means that
you can always see what features are required for a particular item when looking at the docs.

Lastly I also started the switch to support no_std environments, which mainly involved changing a lot of use statements to
now refer to core/alloc instead of std. Then also created the std feature and made it part of the default feature as well as
being enabled by most other features as well. However you should now be able to use most of the Queues in a no_std environment
as long as "alloc" is supported